+++ api/src/main/java/org/bigcraft/infpoints/api/config/PointConfig.java
@@ -2,12 +2,12 @@ package org.bigcraft.infpoints.api.config;
 
 import org.bukkit.configuration.ConfigurationSection;
 
-public record PointConfig(String key, String type, long defaultBalance) {
+public record PointConfig(String key, String type, double defaultBalance) {
 
     public static PointConfig fromConfig(ConfigurationSection section) {
         String key = section.getName();
         String type = section.getString("type");
-        long defaultBalance = section.getLong("defaultBalance");
+        double defaultBalance = section.getLong("defaultBalance");
         return new PointConfig(key, type, defaultBalance);
     }
